Debugging for DSM all the same
Increased the documentation for the whole file.
Added documentation to the set_item() function.
Added the check for the x,y,z co-ordinate in the set.
Introduced n for the type of smk.
changed smk to j in the loop through the smk's.
The bug is caused by line 279. At this step s.d[dk] corresponds to one sparse matrix. But all matrices in the dictionary are set to x.