Fixing Bugs and Speeding Up
Added the indexing options that the last index is a slice.
Sped up get_item by replacing __get_SM_ for a specific matrix tos.d[kd][sm[0],sm[1]] which goes straight to the dictionary element.
Added some sphynx style comments.
Sped up functions by iterating through the nonzero() terms and performing the operations on elements instead of on vectors and sparse matrices. This reduces the if statements and the indexing slowness with the DSM's when going to full columns or rows.
Add the nonzero() indices as inputs to the mul, sub, add, and truediv, getrad, functions. This is an optional input, the shape of ind is checked before iterations started.
Combined multiple function into one forming refcoefdiv which finds the reflectioncoefficients for each term using costhetat and costhetai. refcoefmultdiv computes the replection coefficients from costhetai and the ratios of impedances.
refcoefbyterm computes the reflection coefficients at each term using theta, the impedance ratios and the refractive index.
refcoefbyterm_withmul computes the same as refcoefbyterm which computes the reflection coefficeitns then multiples the terms in the same column.
togrid is sped up by summing all the terms in an sparse matrix rather than taking the norm.
double_dict_vec_multiple multiplies every column in the matrices of a DSM by a vec, it can do this for two DSMs and the same vector.
gain_phase_ref_mul multiplies the terms in a DSM by the combined reflection coefficient terms for parallel and perpendicular refs and the gains.