Fixed bug and Sped up
Fixed the number.r of steps and cone steps.
Sped up the normalisation of the driection vector and the normal vector.
Added the current reflection number to the input for mesh_power_singleray.
Instead of replacing line of sight refs with 0 keep as 1, this is then used to multiply by the output coefficients in the calculation.
Reset i2, j2, k2 not i1, j1, k1.
Corrected the distance correction calculation.