Split as mesh attribute. Add functions to check doubles in a matrix.
Obstacle numbers and reflection numbers now found using the function nob_fromrow() and nre_fromrow().
The doubles in matrix function checks the matrix at an (x,y,z) point and sees if a column corresponding to the current ray has already been stored.
This can be used within the code instead of having to delete doubles later.
check_nonzero_cols is a function for texting. This checks whether the right number of terms is in a column. It either check for a specific stored column or every column in the DSM.