Initialiseation removed from singleray.
In mesh_singleray the initialisation of all the terms used has been removed.
row number is found at the start once then used rather than at the point of storage.
copy is used for p1 and p0 is stored as a reference to the point.
The double in matrix function from DictionarySparseMatrix is used to determine whether the ray has already been stored at that point. The code then skips ray-storing as goes to conestepping.