Update DictionarySparseMatrix.py
Corrected angle check so it looks for when the anlge is wrong and flags an error if so.
Corrected an if to an elif in the reflection LOS and PerfRef check. This was causing a problem when the first statement was satisfied as after the second it was then going to else. This now works for valid LOS and PerfRef values.
Removed the logging info lines at the point (xcheck, ycheck, zcheck).
Added two functions, Watts_to_db and db_to_Watts which converts the way Power is shown.