ResOn, Innerpoints
New parameter ResOn, indicates whether the residual should be calculated.
Inner points are the points which are within the environment but are inside an obstacle. These points are calculated at the start of the code and are an attribute of a room object.
The Radius mesh are no longer output from the power function but saved within the get_rad function.